requirements.txt
PyMuPDF==1.23.26
pytest==7.4.4
pytest-mock==3.12.0
pytest-cov==4.1.0
black==23.12.1
pylint==3.0.3
pyyaml==6.0.1
openai==1.51.0
anthropic==0.34.2
google-generativeai==0.8.3
httpx==0.27.2